Support long coming from db

+++ src/Slp.Evi.Storage/Slp.Evi.Storage.MsSql/QueryWriter/MsSqlQueryWriter.cs
@@ -540,7 +540,7 @@ namespace Slp.Evi.Storage.MsSql.QueryWriter
             }
 
             /// <inheritdoc />
-            public void WriteConstant(int literal)
+            public void WriteConstant(long literal)
             {
                 _sb.Append(literal);
             }
